+++ cg/meta/workflow/mip_dna.py
@@ -1,10 +1,9 @@
 from pathlib import Path
 
-from cg.constants import DEFAULT_CAPTURE_KIT, FileExtensions, Pipeline
-from cg.constants.constants import AnalysisType, FileFormat
+from cg.constants import DEFAULT_CAPTURE_KIT, Pipeline
+from cg.constants.constants import AnalysisType
 from cg.constants.gene_panel import GENOME_BUILD_37
 from cg.constants.pedigree import Pedigree
-from cg.io.controller import WriteFile
 from cg.meta.workflow.mip import MipAnalysisAPI
 from cg.models.cg_config import CGConfig
 from cg.store.models import CaseSample
@@ -76,11 +75,4 @@ class MipDNAAnalysisAPI(MipAnalysisAPI):
         return self._get_managed_variants(genome_build=GENOME_BUILD_37)
 
     def write_managed_variants(self, case_id: str, content: list[str]) -> None:
-        """Write the managed variants to case dir."""
-        out_dir = Path(self.root, case_id)
-        out_dir.mkdir(parents=True, exist_ok=True)
-        WriteFile.write_file_from_content(
-            content="\n".join(content),
-            file_format=FileFormat.TXT,
-            file_path=Path(out_dir, f"managed_variants{FileExtensions.VCF}"),
-        )
+        self._write_managed_variants(out_dir=Path(self.root, case_id), content=content)
